requirements.txt
gitpython==3.1.18
pathlib==1.0.1
python-hcl2==3.0.1
pyyaml==5.4.1
questionary==1.10.0
wheel==0.36.2